pre simulator merge

+++ executor/testdata/cases/py-core/self-run/storage.py
@@ -21,12 +21,12 @@ class Composite:
 
 tst = Composite()
 
-stor_man = tst._storage_slot.manager
+stor_man = tst._storage_slot.manager  # type: ignore
 
-tst.a.foo = 65535
-tst.a.bar = 2**32
+tst.a.foo = i64(65535)
+tst.a.bar = i64(2**32)
 tst.a.st = '123'
-tst.b.foo = 13
+tst.b.foo = i64(13)
 tst.b.st = '321'
 
 stor_man.debug()
